Extract class Children to module
