+++ package.json
@@ -84,8 +84,8 @@
     "bluebird": "3.5.0",
     "classnames": "2.2.5",
     "immutable": "3.8.1",
-    "js-slate-markdown-serializer": "1.0.6",
     "lodash": "4.17.4",
+    "markdown-it": "8.3.1",
     "markdown-it-deflist": "2.0.2",
     "markdown-it-emoji": "1.4.0",
     "markdown-it-footnote": "3.0.1",
