+++ src/client/components/SlateEditor/plugins/slate-format-plugin/EmptyNode.js
@@ -1,7 +1,11 @@
 import React from 'react';
 
+const style = {
+  margin: '0 0 10px',
+};
+
 const EmptyNode = () => {
-  return <div />
+  return <p style={style}> </p>
 };
 
 export default EmptyNode;
